requirements.txt
langgraph>=1.1.0
langchain-openai>=1.1.0
pydantic>=2.8.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pyyaml>=6.0.0
openpyxl>=3.1.0
pandas>=2.2.0
streamlit>=1.40.0
